saving capstone progress

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransfersDAO.java
@@ -1,14 +1,16 @@
 package com.techelevator.tenmo.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
+import com.techelevator.tenmo.model.Accounts;
 import com.techelevator.tenmo.model.Transfers;
 import com.techelevator.tenmo.model.User;
 
 public interface TransfersDAO {
 
 	
-	int getUserCurrentBalanceByID(int userID);
+	BigDecimal getUserCurrentBalanceByID(int userID);
 	
 	List<User> getAllExceptUser();
